api/requirements.txt
aniso8601==9.0.1
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
pymongo==3.11.4
pytz==2021.1
six==1.16.0
Werkzeug==1.0.1
